Update npm scripts to compile and run typescript tests and lint the ts files in the test directory
Update Mocha tests to use TypeScript syntax
